Added `Semaphore`
